refactor/rewrite with es6 goodness
