--- geometry/index.js
@@ -1,4 +1,0 @@
-exports.cylinder = require('./cylinder')
-exports.sphere = require('./sphere')
-exports.plane = require('./plane')
-exports.box = require('./box')
