+++ src/template.js
@@ -1,0 +1,7 @@
+export default `
+  <section class="axis frame">
+    <div class="axis container">
+      <video class="axis" style="display: none; width: 100%; height: 100%;"></video>
+    </div>
+  </section>
+`
